backend/requirements.txt
# Python dependencies

# FastAPI
fastapi
uvicorn
pydantic
python-multipart
python-dotenv

# RAG Stack
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb
groq

# PDF parsing
pypdf

# AWS S3 (for file storage)
boto3

# Database
psycopg2-binary

# Code Quality
black
flake8
isort